improve dir strcuture
